Update the README.md with a new (functional) SLoC badge and add a new COCOMO badge (#175)
This PR updated the README.md with a new (functional) SLoC badge and add a new COCOMO badge.
The current SLoC badge uses token which, way too often, reports 0 or unknown lines of code. On top of this, the current "commits per month" badge doesn't make sense anymore as merges are now getting squashed together to reduce the chaos on the hsitory.
The new SLoC badge works very well.