Fix receive and send when multiple db chunks are involved (#182)
This PR fixes an issue triggered by how streams larger than 64kb are stored in the database and also adjust the size of the buffers.
The PR also improves how data are sent when multiple chunks have to be streamed, now the chunks are properly split depending on the size of the send buffer instead of requiring the send buffer to be the size of the data to be sent.

Another fix included in the PR is related to properly handling short sends, as these are legit and were causing tests to fail when sending too much data. In addition a check has been added to ensure that when the slab allocator is enabled the maximum length of the allowed arguments to be saved in the storage db doesn't exceed ~170MB (64kb / 24b [size of the chunk info struct] * 64kb).