Implement a queue to track the epoch of the requested and completed operations (#257)
The PR implements an operation queue to track the epoch of the
operations and when they are completed.

This queue will be used by the workers to identify up to which point the
epoch can be moved forward for the single workers and will be used in
combination with the epoch gc.

Each time an operation is started an operation will be registered in the
queue and when the operation will finish it will be marked as completed.

The worker then, randomly, will process the entities at the beginning of
this queue to determine how many operations have completed and how much
the queue can be moved forward.

The epoch operation queue uses a specialized 64bit ring unbounded queue
using the MSB bit (bit n. 64) to indicate the completition and the
remaining 63 bits to store the epoch, a 63 bits epoch is able to count
about 100 years of runtime on a 3ghz cpu, should be plenty for the
forseeable future.