Add prometheus stats per worker (#269)
To properly analyze and investigate the load the prometheus module now
expose all the standard metrics per worker, not only the aggregated
ones.

To differenciate between the metrics per worker and the aggregated ones
a tag is added per metric to indicate the worker index which is set to
"aggregated" for the aggregated metrics.

This change breaks compatibility with the previous behaviour and not
filtering on the aggregated stats will cause to any software reading the
prometheus metrics to provide incorrect values. The new setups should
take this into account as well.