Initial implementation transactions in new hastable mpmc (#285)
This PR introduces a basic transaction support for the new mpmc
hashtable:
- it introduces two new data structures, both mpmc and spsc, to fetch
and release slot ids using a bitmap, whereas each slot is mapped to each
bit of the bitmap
- this special bitmap uses a table lookup to quickly find which is the
first bit available in an uint16_t drammatically reducing the number of
comparisons the algorithm has to carry out
- the mpmc variant offers some additional functions over the spsc one to
avoid cache-conflicts in the vast majority of the cases
- it restructures the transactions related structs and fields, now the
id is a plain uint32_t number
- it can even be reduced in size to recoup space to be used for
different purposes, probably a 20 bit number would be plenty as having
1048575 transactions seems unlikely even on very high end systems
- it updates the internal interfaces to provide the needed hooks and
checks for the transactions needed for the Read-Modify-Write operations

This PR is extremely important because it introduces a blazing fast new
data structure, the `slots_bitmap_mpmc` which performs extremely well
under contention. The main idea is that each transaction is just a slot
of the bitmap and each thread will search in this bitmap first in its
own shard (as the bitmap allows localized searches) and then in the
whole bitmaps as fallback.

The result is that in a normal situation, where slots are available for
the thread, to acquire and release 51200 slots per threads it takes just
a few nanoseconds (more details in
`benches/bench-slots-bitmap-mpmc.cpp`)
```
threads:1 15099228 ns
threads:2 1027204 ns
threads:4 74470 ns
threads:8 5860 ns
threads:16 499 ns
threads:32 42 ns
```

With 4 thread and in avg 800k ops/s in total (therefore potentially 800k
transactions/s) the time required would be 1.16ms, an average overhead
of 1.45ns per transaction to be acquired and released meanwhile with
with just 1 thread and in avg 200k ops/s would be 295us.
In both cases the time to acquire and release a transaction id is close
enough to zero to be considered irrelevant!