Upgrade Catch2 to v3.3.2 with Compact Headers (#291)
This pull request upgrades Catch2 to version v3.3.2, incorporating the
use of compact headers and updating tests to utilize the latest version
of Catch. The key changes include switching to compact headers for
improved readability and maintainability, updating Catch2 to v3.3.2 for
enhanced performance and bug fixes, and updating new tests to ensure
compatibility with the upgraded Catch2 library.