Bug fix listeners client channel memory leak (#348)
The listeners were always setting up a new channel before accepting a
new connection to be able to write the socket address, and the socket
address truct length, directly in the channel.

The side effect though is that when a listener is terminated that memory
is lost.

To avoid an useless memory leak, the channel is now allocated only if
there is a new connection and the socket address as well the socket
address length is copied. Although this is an additional memory copy,
the amount of data transferred is small enough to do not be significant.
Also it happens only at connection time where there are latencies caused
by other components, not even dependent on cachegrand (e.g. dns
resolution, tcp handshake, etc.).