Implement Redis authentication (#396)
This PR implements an authentication mechanism for Redis, with a number
of commands configured to require it if enabled.

3 new parameters have been introduced:
- `require_authentication`, boolean
- `username`, string
- `password`, string

If `require_authentication` is set to true, `password` is mandatory,
`username` is never mandatory and if not specified defaults to `default`
as per Redis implementation. If `require_authentication` is set to
false, a warning will be printed out if `username` or `password` contain
something.

The PR also implements the AUTH command and update the HELLO command to
support the authentication.

The tests have been switched to use a built from code hiredis library,
the one provided by the OS is way too old and doesn't support RESP3.
The tests for the HELLO command have been updated to use the internal
mechanism to send Redis commands via hiredis instead of sending and
decoding the data manually.