Zero memory when upsize args list (#411)
This PR fixes a bug caused by a previously implemented optimizaiton when
the ffma_mem_realloc interface was changed to match the realloc
interface.

Previously, the function was taking a parameter to zero the newly
allocated memory, after the change that was moved outside and an if to
zero the memory only when the location of the pointer was changing was
introduced.

This introduced the current bug as the zero-ed memory previously was
always the right amount but as now the operation is performed outside
ffma only the memory to be used was being zeroed. The code changes that
behaviour simply zeroing all the memory to be used always, it doesn't
introduce any additional overhead as the zero-ing was previously done
before the ffma_mem_realloc changes.

A test to catch this issue with mset has been introduced as well.