Add support for the PING command over the inline protocol (#412)
This PR adds, as a quick workaround to fully enable the redis-benchmark
cli tool support, the ability to handle the PING command over the inline
protocol.

The implementation is fairly simple and works only if the PING command,
followed by the new line, fits, in its entirety, the buffer otherwise is
reported as error and the connection is closed. As the only reason of
this implementation is to support redis-benchmark and it never sends
commands byte by byte, we can always expect the command to be in the
buffer.

The PR also contains some minor valgrind and comment fixes.