Rely on read tx for the hashtable consistency (#434)
Now that every operation on the hashtable is always behind a read or
write transaction, it's not necessary to have to rely on a tricky logic
and memory fences to ensure that a read will never cause an issue if a
write changes something as they will never run in parallel.

The PR drops the memory fences, introduces some minor optimizations now
that reads are consistent and clean up a lot of unnecessary logging.

The PR also contains a minor fix for a build warning when building C++
code (e.g. benchmarks or tests) and the necessary changes to update the
set benchmark for the hashtable to the new paradigm implemented.