Drop unused new ht and epoch gc (#437)
This PR drops the more recent implementation of the faster hashtable,
which although it's faster, it's not built to support transactions and,
even if they would be "forced" into the architecture as the only goal of
this implementation was to allow parallel access having the transaction
would negate this core benefit.

For this reason I decided to drop all the code related to the new and
faster hashtable as well as the epoch gc support. There are other bits
that were implemented to support these two components that aren't
removed as they are small things that can be useful in the future.