Add build for ubuntu 24.04 when merging PRs (#446)
This pull request includes a new job configuration for Ubuntu 24.04 in
the build and test workflow, testing out both GCC13 and GCC14. It also drops the
lcov github action in faviour of invoking lcov directly within the pipeline.

*
[`.github/workflows/build_and_test.yml`](diffhunk://#diff-48c0ee97c53013d18d6bbae44648f7fab9af2e0bf5b0dc1ca761e18ec5c478f2R27-R33):
Added a new job configuration for "Ubuntu 24.04" with `os` set to
`ubuntu-24.04`, `triplet` set to `x64-linux`, and compilers `gcc` and
`g++`.