Feat/update ubuntu build documentation (#447)
This pull request includes updates to the documentation for building
from source. The most important changes are the consolidation of Ubuntu
installation instructions and the removal of outdated instructions for
Ubuntu 20.04.

Documentation updates:

*
[`docs/build-from-source.md`](diffhunk://#diff-98e27fe46505c0c406242be6aafa31c1e882a6206812438ff870db3075e14d7bL34-R34):
Updated the section title to include both Ubuntu 22.04 and Ubuntu 24.04.
*
[`docs/build-from-source.md`](diffhunk://#diff-98e27fe46505c0c406242be6aafa31c1e882a6206812438ff870db3075e14d7bL52-L69):
Removed the outdated installation instructions for Ubuntu 20.04.